requirements.txt
ftfy==6.3.1
numpy==2.2.6
Pillow==12.1.1
PyYAML==6.0.3
regex==2026.2.28
scikit-learn==1.7.2
scipy==1.15.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.3